requirements.txt
--pre 
azure-ai-projects>=2.0.0b1
azure-identity
openai
python-dotenv